Add `riscv64gc-unknown-linux-gnu` as default target
This is because more machines with RISC-V CPUs are arriving, and they are
good for Nerves projects, since they are cheaper then Raspberry Pis.

In this commit there is also a change to prevent duplicated values,
since people may already added some of the targets that are in the list
of defaults.

Closes https://github.com/philss/rustler_precompiled/issues/44