Make sure `:nif_versions` option is respect (#60)
This is a small bug fix that was blocking the usage of a more
restrict list of "NIF versions". For example, if my system is
using NIF version 2.16, but I want to be compatible with only
version 2.15, this was being ignored, since the algorithm for
finding compatible versions was not taking into account this
restriction.