Add support for setting a global cache path using an env var
This is a feature that can help people with internet access restrictions
or writing permission limitations.

It may help people that need to support systems running on NixOS or a
Nix environment.

In a nutshell, first download the desired ".tar.gz" artifact for a given
package, set the `RUSTLER_PRECOMPILED_GLOBAL_CACHE_PATH` to the
directory that the package was downloaded, and then build your project
normally.